InProgress #35851 - [Fix] Check nodes existence before deleting them in alfrescoRepository

+++ becpg-core/src/main/java/fr/becpg/repo/repository/impl/AlfrescoRepositoryImpl.java
@@ -608,7 +608,7 @@ public class AlfrescoRepositoryImpl<T extends RepositoryEntity> implements Alfre
 					boolean deleteNodes = false;
 
 					for (RepositoryEntity dataListItem : ((LazyLoadingDataList<? extends RepositoryEntity>) dataList).getDeletedNodes()) {
-						if ((dataListItem != null) && (dataListItem.getNodeRef() != null) && !dataListItem.isTransient()) {
+						if ((dataListItem != null) && (dataListItem.getNodeRef() != null) && !dataListItem.isTransient() && nodeService.exists(dataListItem.getNodeRef())) {
 							nodeService.addAspect(dataListItem.getNodeRef(), ContentModel.ASPECT_TEMPORARY, null);
 							nodeService.deleteNode(dataListItem.getNodeRef());
 							deleteNodes = true;
